Partial diaspy implementation
Began converting the gather() method for DiasporaTools to diaspy. So far,
this can harvest a user's activities but it doesn't check primary
authorship. It does set the post date correctly and sets the link and ID.
More work needed. Have to check author and establish that the author
is the user, or if not that this is a share/reshare.